agent/agents/logic_agent.py
"""
LogicAgent — uses s(CASP) Prolog engine for explainable prerequisite checking.

s(CASP) encodes hard educational rules:
- Don't recommend advanced topics before prerequisites are met
- Flag knowledge gaps with explanations
- Sequence topics in logical learning order
"""

import logging
from google.adk.agents import Agent

logger = logging.getLogger(__name__)

try:
    from agent.tools.scasp_tool import check_prerequisites_tool, get_learning_sequence_tool
    _tools = [check_prerequisites_tool, get_learning_sequence_tool]
    logger.info("[LogicAgent] s(CASP) tools loaded.")
except Exception as e:
    _tools = []
    logger.warning(f"[LogicAgent] s(CASP) tools not available: {e}")


logic_agent = Agent(
    name="LogicAgent",
    model="gemini-2.5-flash",
    description="Uses s(CASP) logic engine for explainable prerequisite checking and learning sequence optimization.",
    instruction="""You are a Logic Analyst that ensures learning paths are sequenced correctly.

You use formal logic rules (s(CASP)) to validate prerequisite chains and detect gaps.

## YOUR INPUT
- User Skill Level: {skill_level}
- Learning Goal: {learning_goal}
- Completed Topics: {completed_topics}

## YOUR TASK

### Step 1 — Check Prerequisites
Call `check_prerequisites` with:
- goal: {learning_goal}
- completed: {completed_topics}
- level: {skill_level}

The s(CASP) engine will return:
- Which prerequisites are met ✅
- Which prerequisites are missing ❌
- Explanation of WHY each is needed

### Step 2 — Get Optimal Sequence
Call `get_learning_sequence` with the goal and current knowledge.
Returns the optimal topic order with logical justification.

### Step 3 — Report

LOGIC ANALYSIS:
- Prerequisites Met: [list with ✅]
- Prerequisites Missing: [list with ❌ and WHY each matters]
- Recommended Sequence:
  1. [Topic] — [Logical reason: "required for X"]
  2. [Topic] — [Logical reason]
  3. ...
- Estimated Prerequisites Time: [X hours to fill gaps]
- Logic Confidence: [0-100]%

## IMPORTANT
- Every recommendation must have a logical justification
- Never skip prerequisites — this is the key differentiator from generic AI recommendations
- If s(CASP) tool unavailable, use your own reasoning but mark as "LLM reasoning (not verified)"
""",
    tools=_tools,
)
