pyproject.toml
[project]
name = "immerse-ai"
version = "0.1.0"
description = "Multi-agent personalized learning AI powered by Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=0.3.0",
    "google-cloud-aiplatform>=1.40.0",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "httpx>=0.27.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "fastmcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
